New version comparaison algorithm
Instead of using LooseVersion from distutils (a missed deps btw), I switched to
a custom version not directly based on tuples comparaison. Because it doesn't
support different type inside tuple.

All my package are currently passing better than with LooseVersion.

We also doesn't compute the full sorted list of versions but only the max, which
is the last.