Retrive first pkgname insteaf of pkgbase by default
To keep the rules of using binary name of the package in the config file,
we grab the pkgname[0] of the bash array. This is done to stay compatible with
archweb.

As bash 4.2.x doesn't export array to env (well know bug). We add a custom
export as string the first field of pkgname array as pkgname0.